+++ src/main/java/org/example/rowmapper/LandmarkFullRowMapper.java
@@ -1,6 +1,5 @@
 package org.example.rowmapper;
 
-import org.example.model.LandmarkBasicModel;
 import org.example.model.LandmarkFullModel;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
@@ -9,6 +8,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 // Полный
 @Component
+
 public class LandmarkFullRowMapper implements  RowMapper<LandmarkFullModel> {
     @Override
 public LandmarkFullModel mapRow(ResultSet rs, int rowNum) throws  SQLException {
@@ -24,8 +24,8 @@ public LandmarkFullModel mapRow(ResultSet rs, int rowNum) throws  SQLException {
                 rs.getString("open"),
                 rs.getString("close"),
                 rs.getBoolean("available"),
-                rs.getDouble("lat"),
-                rs.getDouble("lon"),
+                rs.getFloat("lat"),
+                rs.getFloat("lon"),
                 rs.getString("image")
         );
     }
